+++ tests/parser_integration.rs
@@ -53,9 +53,49 @@ fn test_get_ast_with_backreference() {
             UnitExpression::Primitives {
                 data_type: DataType::String(
                     PositiveReferenceType::ByLiteral(clex_language::ast::MAX_STRING_SIZE as u64),
-                    CharacterSet::default_charset()
+                    CharacterSet::get_default_charset()
                 ),
                 repetition: PositiveReferenceType::ByGroup { group_number: 1 },
+            },
+            UnitExpression::Eof
+        ]
+    );
+}
+
+#[test]
+fn test_custom_charset_with_string() {
+    let language = "S[10, 'asghdgad']";
+    let ast = get_ast(language.to_string()).unwrap();
+
+    assert_eq!(
+        ast.expression,
+        vec![
+            UnitExpression::Primitives {
+                data_type: DataType::String(
+                    PositiveReferenceType::ByLiteral(10),
+                    CharacterSet::Custom("asghdgad".to_string())
+                ),
+                repetition: PositiveReferenceType::ByLiteral(1),
+            },
+            UnitExpression::Eof
+        ]
+    );
+}
+
+#[test]
+fn test_charset_with_string() {
+    let language = "S[10, @CH_ALL@]";
+    let ast = get_ast(language.to_string()).unwrap();
+
+    assert_eq!(
+        ast.expression,
+        vec![
+            UnitExpression::Primitives {
+                data_type: DataType::String(
+                    PositiveReferenceType::ByLiteral(10),
+                    CharacterSet::All
+                ),
+                repetition: PositiveReferenceType::ByLiteral(1),
             },
             UnitExpression::Eof
         ]